+++ tasks/wrappers/aws.js
@@ -7,11 +7,22 @@ var fs = require('../utils/fs');
 function AWS(location, destination, options){
     this.location = location;
     this.destination = this.addSlash(destination);
-    this.config = {
-        accessKey: options.accessKey || process.env.AWS_ACCESS_KEY_ID || null,
-        secret: options.secret || process.env.AWS_SECRET_ACCESS_KEY || null,
-        region: options.region || process.env.AWS_REGION || null
-    };
+    AWSSDK.config.update({region: options.region || process.env.AWS_REGION || null});
+    var credentials;
+    if (options.profile) {
+        credentials = new AWSSDK.SharedIniFileCredentials({profile: options.profile});
+    }
+    var auth;
+    if (credentials) {
+        auth = { credentials: credentials };
+    } else if (options.accessKey && options.secret) {
+        auth = {
+            accessKeyId: options.accessKey,
+            secretAccessKey: options.secret
+        };
+    }
+    this.s3 = new AWSSDK.S3(auth);
+
     this.params = {
         Bucket : options.bucket || process.env.AWS_BUCKET || null,
         ACL    : options.acl || 'public-read',
@@ -26,15 +37,12 @@ AWS.prototype.addSlash = function(dir){
 };
 
 AWS.prototype.checkMandatory = function(key){
-    if (!this.config[key] && !this.params[key]) {
+    if (!this.params[key]) {
         log.onError({message:'AWS: Missing config `' + key + '`'});
     }
 };
 
 AWS.prototype.setParams = function(fileObj){
-    this.checkMandatory('accessKey');
-    this.checkMandatory('secret');
-    this.checkMandatory('region');
     this.checkMandatory('Bucket');
 
     if (fileObj.ext) {
@@ -53,13 +61,7 @@ AWS.prototype.upload = function(fileObj) {
     var self = this;
     this.setParams(fileObj);
     return new Promise(function(resolve, reject){
-        var s3 = new AWSSDK.S3({
-            accessKeyId     : self.config.accessKey,
-            secretAccessKey : self.config.secret,
-            region          : self.config.region
-        });
-
-        s3.putObject(self.params, function(err) {
+        self.s3.putObject(self.params, function(err) {
             if (err) {
                 reject({message: 'S3::putObject "' + self.params.Key + '" error!\n' + err});
             } else {